Security Update
BE : Am adaugat BCrypting la parola. Si merge cea ce este fascinant

FE : Am adaugat campul "confirm password".

+++ ProbaIt FE/signup.js
@@ -8,35 +8,40 @@ document.addEventListener("DOMContentLoaded", function() {
         const username = document.getElementById("username").value;
         const email = document.getElementById("email").value;
         const password = document.getElementById("password").value;
+        const confirmPassword = document.getElementById("confirmPassword").value;
 
-        // Prepare JSON body
-        const requestBody = {
-            "username": username,
-            "email": email,
-            "password": password
-        };
+        if(password == confirmPassword){
+            // Prepare JSON body
+            const requestBody = {
+                "username": username,
+                "email": email,
+                "password": password
+            };
 
-        // Send POST request
-        fetch("http://localhost:8080/signup", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify(requestBody)
-        })
-        .then(response => {
-            if (!response.ok) {
-                throw new Error("Signup failed");
-            }
-            return response.json();
-        })
-        .then(data => {
-            localStorage.setItem('loggedUserId', data.id);
-            localStorage.setItem('loggedUserUsername', data.username);
-            window.location.href = "landing page.html";
-        })
-        .catch(error => {
-            console.error("Signup error:", error);
-        });
+            // Send POST request
+            fetch("http://localhost:8080/signup", {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify(requestBody)
+            })
+            .then(response => {
+                if (!response.ok) {
+                    throw new Error("Signup failed");
+                }
+                return response.json();
+            })
+            .then(data => {
+                localStorage.setItem('loggedUserId', data.id);
+                localStorage.setItem('loggedUserUsername', data.username);
+                window.location.href = "landing page.html";
+            })
+            .catch(error => {
+                console.error("Signup error:", error);
+            });
+        }else {
+            alert('Passwords don\'t match. (nu suna misto in romana)')
+        }
     });
 });
